aga/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aga-core"
version = "4.4.0"
description = "AGA — Auxiliary Governed Attention: 极简注意力治理插件，为冻结 LLM 提供推理时知识注入能力"
readme = "aga/README_en.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AGA Team"},
]
keywords = ["llm", "attention", "knowledge-injection", "inference", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
knowledge = ["aga-knowledge>=0.3.0"]
observability = ["aga-observability>=1.0.0"]
all = [
    "pyyaml>=6.0",
    "aga-knowledge>=0.3.0",
    "aga-observability>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/aga-project/aga-plugin"
Documentation = "https://aga-project.github.io/docs"
Repository = "https://github.com/aga-project/aga-plugin"

[tool.setuptools.packages.find]
include = ["aga*"]
exclude = ["aga_knowledge*", "aga_observability*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
